armv7 tests: try to cache rpi2 toolchain

+++ scripts/ci/linuxarmv7/install.sh
@@ -28,10 +28,11 @@ downloadToolchain(){
     #tar xf x-tools7h.tar.xz
     #rm x-tools7h.tar.xz
     
-    
-    wget http://ci.openframeworks.cc/rpi2_toolchain.tar.bz2
-    tar xjf rpi2_toolchain.tar.bz2
-    rm rpi2_toolchain.tar.bz2
+    if [ ! -d ~/rpi2-toolchain ]; then
+        wget http://ci.openframeworks.cc/rpi2_toolchain.tar.bz2
+        tar xjf rpi2_toolchain.tar.bz2 -C ~/
+        rm rpi2_toolchain.tar.bz2
+    fi
 }
 
 downloadFirmware(){
